cocktails images

+++ terranova.Server/Services/AzureStorageService.cs
@@ -7,12 +7,15 @@ namespace terranova.Server.Services
     {
         Task<string> UploadProfileImageAsync(string userId, Stream imageStream, string contentType);
         Task<bool> DeleteProfileImageAsync(string userId, string imageUrl, bool isAdmin = false);
+        Task<string> UploadCocktailImageAsync(string userId, Stream imageStream, string contentType);
+        Task<bool> DeleteCocktailImageAsync(string userId, string imageUrl, bool isAdmin = false);
     }
 
     public class AzureStorageService : IAzureStorageService
     {
         private readonly BlobServiceClient _blobServiceClient;
         private readonly string _containerName = "profile-images";
+        private readonly string _cocktailContainerName = "cocktail-images";
 
         public AzureStorageService(IConfiguration configuration)
         {
@@ -38,7 +41,7 @@ namespace terranova.Server.Services
                 return false;
 
             var blobName = Path.GetFileName(uri.LocalPath);
-            string expectedPrefix = $"{userId}-";
+            string expectedPrefix = $"propic-{userId}-";
             if (!isAdmin)
             { 
                 if (!blobName.StartsWith(expectedPrefix))
@@ -48,6 +51,39 @@ namespace terranova.Server.Services
             }
 
             var containerClient = _blobServiceClient.GetBlobContainerClient(_containerName);
+            var blobClient = containerClient.GetBlobClient(blobName);
+
+            return await blobClient.DeleteIfExistsAsync();
+        }
+
+        public async Task<string> UploadCocktailImageAsync(string userId, Stream imageStream, string contentType)
+        {
+            var containerClient = _blobServiceClient.GetBlobContainerClient(_cocktailContainerName);
+            await containerClient.CreateIfNotExistsAsync(PublicAccessType.Blob);
+
+            var blobName = $"cocktail-{userId}-{Guid.NewGuid()}";
+            var blobClient = containerClient.GetBlobClient(blobName);
+
+            await blobClient.UploadAsync(imageStream, new BlobHttpHeaders { ContentType = contentType });
+            return blobClient.Uri.ToString();
+        }
+
+        public async Task<bool> DeleteCocktailImageAsync(string userId, string imageUrl, bool isAdmin = false)
+        {
+            if (string.IsNullOrEmpty(userId) || !Uri.TryCreate(imageUrl, UriKind.Absolute, out Uri uri))
+                return false;
+
+            var blobName = Path.GetFileName(uri.LocalPath);
+            string expectedPrefix = $"cocktail-{userId}-";
+            if (!isAdmin)
+            {
+                if (!blobName.StartsWith(expectedPrefix))
+                {
+                    return false;
+                }
+            }
+
+            var containerClient = _blobServiceClient.GetBlobContainerClient(_cocktailContainerName);
             var blobClient = containerClient.GetBlobClient(blobName);
 
             return await blobClient.DeleteIfExistsAsync();
